used size_t instead of int at token_length of GLSLTokenizer.h

+++ GLSLTokenizer.h
@@ -10,7 +10,7 @@ namespace shader_lib
         int source_length;
 
         GLSLToken token_type;
-        int token_length;
+        size_t token_length;
 
 	private:
